moved create_faceted_dotplot_with_percentages from analysis to data_exploration
